fix: 优化 useStorage
